+++ SkyModel/PEX.py
@@ -268,7 +268,7 @@ def main(options=None):
     ylist=[]
     slist=[]
 
-    Cat=np.zeros((50000,),dtype=[('ra',np.float),('dec',np.float),('s',np.float),('I',np.float),('Gmaj',np.float),('Gmin',np.float),('PA',np.float),('Gangle',np.float)])
+    Cat=np.zeros((50000,),dtype=[('ra',float),('dec',float),('s',float),('I',float),('Gmaj',float),('Gmin',float),('PA',float),('Gangle',float)])
     Cat=Cat.view(np.recarray)
 
     isource=0
