+++ SkyModel/Sky/ClassClusterClean.py
@@ -229,7 +229,7 @@ class ClassClusterClean():
             x0,x1,y0,y1=extent
         gx,gy=np.mgrid[x0:x1:Ncells*1j,y0:y1:Ncells*1j]
         
-        CatCell=np.zeros((gx.flatten().shape[0],),dtype=[("ToNumNode",np.int),("xcell",float),("ycell",float)])
+        CatCell=np.zeros((gx.flatten().shape[0],),dtype=[("ToNumNode",int),("xcell",float),("ycell",float)])
         CatCell=CatCell.view(np.recarray)
         CatCell.xcell=gx.reshape((NcellsSq,))
         CatCell.ycell=gy.reshape((NcellsSq,))
